backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
httpx
motor 
groq
email-validator
python-jose
passlib
python-multipart
bcrypt